Refactor member variable transformer
This removes the member variable transformer logic from
ClassTransformer. Instead we use the alrady existing
ES6ClassTransformer which does this as a stand alone transformer.
