+++ src/codegeneration/MemberVariableConstructorTransformer.js
@@ -23,7 +23,6 @@ import {
   createThisExpression,
 } from './ParseTreeFactory.js';
 import {prependStatements} from './PrependStatements.js';
-import {parseExpression} from './PlaceholderParser.js';
 
 /**
  * Transforms class constructors when classes have initialized instance
@@ -49,15 +48,6 @@ export function transformConstructor(constructor, initExpression, superClass) {
       constructor.functionKind, constructor.name, constructor.parameterList,
       constructor.typeAnnotation, constructor.annotations,
       createFunctionBody(statements), constructor.debugName);
-}
-
-// TODO(vicb): Does not handle computed properties
-export function getInstanceInitExpression(initInstanceVars) {
-  let expressions = initInstanceVars.map((mv) => {
-    let name = mv.name.literalToken;
-    return parseExpression `this.${name} = ${mv.initializer}`;
-  });
-  return createCommaExpression(expressions);
 }
 
 class SuperCallTransformer extends ParseTreeTransformer {
